Changed Attacks to interface-based implementation
